% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltanets.R
\name{deltanets}
\alias{deltanets}
\title{DeltaNeTS}
\usage{
deltanets(lfc, slope = NULL, glist = NULL, grn = NULL, kfold = 10,
  par = FALSE, numCores = 4)
}
\arguments{
\item{lfc}{The numeric matrix or data.frame of log2FC data. Each row represents a gene and each column represents a sample.}

\item{slope}{The slope matrix from log2FC data. This matrix can be obtained using the function \code{generateSlope()}. If the data are not time-series, set slope to an empty matrix (i.e. \code{slope}=\code{NULL}).}

\item{glist}{The list of genes following the same order as the one in \code{lfc}. If glist is \code{NULL}, the default is just a numbering from 1 to n.}

\item{grn}{The list of TF-gene interactions (optional). The gene symbol in \code{grn} should be the same style of the one in \code{glist}. If glist is not available, use indices in \code{grn} instead of gene symbol. If grn is used, ridge regression is used. Otherwise, lasso regression is used to infer A and P.}

\item{kfold}{The number of folds used in the k-fold cross validation.}

\item{par}{A Boolean variable \code{TRUE} or \code{FALSE} indicating whether to use parallel computing. The default is \code{FALSE} (no parallel computation).}

\item{numCores}{The number of CPU cores to be used for parallel computing. This parameter is considered only if par is \code{TRUE}. The default is 4.}
}
\value{
a list of two matrices
\item{P}{The matrix of gene perturbations. Each row corresponds to a gene following the same order as the one in the log2FC data, while each column corresponds to samples as defined in the tobject.}
\item{A}{The n*n matrix of the inferred gene regulatory network. The (i,j)-th element of the matrix corresponds to the regulatory impact of gene j on gene i. The rwos and columns of the matrix correspond to genes following the same order as the one in the log2FC data.}
}
\description{
The main function for DeltaNeTS for inferring gene perturbations for each sample.
}
